PMM-6067 replace time range change logic
